+++ test/test.js
@@ -7,6 +7,10 @@ describe('etag(entity)', function () {
     it('should generate an ETag', function () {
       assert.equal(etag('beep boop'), '"677e121b2436201ed8cc1ec73240a3ad"')
     })
+
+    it('should work containing Unicode', function () {
+      assert.equal(etag('论'), '"696f4778b4e4d98b7a95fef325812583"')
+    })
   })
 
   describe('when "entity" is a Buffer', function () {