+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix behavior of string containing Unicode
+
 1.0.0 / 2014-05-18
 ==================